requirements.txt
aiohttp==3.4.4
amqp==2.3.2
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
billiard==3.5.0.4
celery==4.0.2
certifi==2018.4.16
chardet==3.0.4
coinmarketcap==5.0.3
crypto-news-api==2.3.1
Cython==0.28.5
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.37.0
django-appconf==1.0.2
django-bootstrap-form==3.4
django-debug-toolbar==1.9.1
django-extensions==1.7.3
django-imagekit==4.0.2
django-render-block==0.5
enum-compat==0.0.2
future==0.16.0
idna==2.7
idna-ssl==1.1.0
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
kombu==4.1.0
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
multidict==4.4.0
oauthlib==2.1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pilkit==2.0
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2018.5
rc-django-celery-results==1.0.1
redis==2.10.6
requests==2.19.1
requests-cache==0.4.13
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.11.11
wrapt==1.10.11
yarl==1.2.0
